move types to '@frontity/core/types'
