+++ packages/core/src/config/frontity/index.ts
@@ -1,4 +1,4 @@
-import { FrontityConfig } from "../../types";
+import { FrontityConfig } from "../../../types";
 
 export default (): FrontityConfig => ({
   outDir: "build"
