+++ packages/core/src/config/webpack/externals.ts
@@ -1,5 +1,5 @@
 import { Configuration } from "webpack";
-import { Target } from "../../types";
+import { Target } from "../../../types";
 
 export default ({ target }: { target: Target }): Configuration["externals"] => {
   const config: Configuration["externals"] = {};
