+++ packages/core/src/config/webpack/name.ts
@@ -1,5 +1,5 @@
 import { Configuration } from "webpack";
-import { Target } from "../../types";
+import { Target } from "../../../types";
 
 // It is important to use the names 'client' and 'server' for
 // webpack-hot-server-middleware.
