+++ packages/core/src/scripts/dev.ts
@@ -8,7 +8,7 @@ import HotServer from "./utils/hot-server";
 import generateEntryPoints from "./utils/entry-points";
 import getConfig from "../config";
 import getFrontity from "../config/frontity";
-import { Mode } from "../types";
+import { Mode } from "../../types";
 import cleanBuildFolders from "./utils/clean-build-folders";
 import { webpackAsync } from "./utils/webpack";
