+++ packages/core/src/server/templates/index.ts
@@ -1,6 +1,6 @@
 import html from "./html";
 import amp from "./amp";
-import { Template } from "../../types";
+import { Template } from "../../../types";
 
 export default ({ mode }: { mode: string }): Template =>
   mode === "amp" ? amp : html;
