+++ packages/core/typings.d.ts
